Revert "stuff1"
This reverts commit 4c1786aa436cb53c93ba378480d59348f0f3c610.

+++ src/async_impl/i2c_helper.rs
@@ -23,7 +23,7 @@ pub(crate) struct I2CHelper<I2C, D> {
     i2c_interface: I2C,
     address: u8,
     delayer: D,
-    pub ambient_temperature: i8,
+    pub ambient_temperature: i32,
 }
 impl<I2C, D> I2CHelper<I2C, D>
 where
@@ -34,7 +34,7 @@ where
         i2c_interface: I2C,
         device_address: DeviceAddress,
         delayer: D,
-        ambient_temperature: i8,
+        ambient_temperature: i32,
     ) -> Self {
         Self {
             i2c_interface,
